added hedgedoc_url to challenges
When a challenge is created, a url should be created with it(didn't end up testing the actual creation of challenge because of errors in ctf creation, but the function to create the url(in this case just the note) works, as i had written a test for it, and tested it. To have hedgedoc up and running use the .env file provided in the readme file.