Made it so that it stores token in cookies instead
There are some errors happening when submitting something in the form, this error is:
{
    "message": "Cross-site POST form submissions are forbidden"
}

which i think is happening because of SvelteKit's CSRF safeguard. We should, if possible, serve our SvelteKit app and our API from the same domain (and port) maybe using ngix. That way, the request is considered same‑site and the CSRF check won’t block it. This isn't as important for now i believe because things still seem to work and show when someone creates a CTF event room and challenges.