Added real time updates to admin dashboard
There is an error about showing the wrong event (shows the newChallenge in the ctfs.go in the createCTF endpoint), when refreshing it fetches the correct event. Still works, just need to do a little bit of modification. Will try to add on to have real time updates for challenges table as well.