Added some functionality to active members in the Event Page
Added table to store participants in CTF, created routes to add, remove and get the participants, and changed/added to event-room page to show a working members list. There are still some bugs needing to be fixed, like an issue with closing the page when you have 2 windows open, it removed the participant, tried to work to fix it by adding a form of page tab count, but it also has bugs, will try to see if there is a better method to do so. Also made it so that if challenges were added/deleted it would also happen in challenge page for the challenge drop down. Have not implemented participant tracking in challenge page.